no imagination, no wings

+++ 0x0B-malloc_free/main.h
@@ -5,7 +5,7 @@
 #include <stdlib.h>
 
 char *create_array(size_t size, char c);
-int _strlen(char *str);
+size_t _strlen(char *str);
 char *_strdup(char *str);
 
 #endif
